+++ src/observer.rs
@@ -46,12 +46,12 @@
 mod observers;
 
 use crate::{
+    cam::{CieCam02, CieCam16, ViewConditions},
     error::Error,
     illuminant::{CieIlluminant, Planck},
     lab::CieLab,
     rgb::RgbSpace,
-    spectrum::to_wavelength,
-    spectrum::{Spectrum, NS, SPECTRUM_WAVELENGTH_RANGE},
+    spectrum::{to_wavelength, Spectrum, NS, SPECTRUM_WAVELENGTH_RANGE},
     traits::{Filter, Light},
     xyz::XYZ,
 };
@@ -157,6 +157,14 @@ impl Observer {
         self.data().name
     }
 
+    fn xyz_and_xyzn(&self, light: &dyn Light, filter: &dyn Filter) -> [XYZ; 2] {
+        let xyzn = light.xyzn(self.data().tag, None);
+        let s = *light.spectrum() * *filter.spectrum();
+        let xyz = self.xyz_from_spectrum(&s);
+        let scale = 100.0 / xyzn.xyz.y;
+        [xyz * scale, xyzn * scale]
+    }
+
     /// Calulates Tristimulus values for an object implementing the [Light] trait, and an optional [Filter],
     /// filtering the light.
     ///
@@ -178,14 +186,72 @@ impl Observer {
         }
     }
 
-    /**
-        Calculates Tristimulus valus, in form of an [XYZ] object of a general spectrum.
-        If a reference white is given (rhs), it will copy its tristimulus value, and the spectrum
-        is interpreted as a stimulus, being a combination of an illuminant with a colorant.
-        If no reference white is given, the spectrum is interpreted as an illuminant.
-        This method produces the raw XYZ data, not normalized to 100.0
+    /// Calculates the L*a*b* CIELAB values for a light source and filter combination.
+    /// This method is used to compute the color appearance of a light source
+    /// when filtered by a colorant or filter.
+    /// It returns the CIELAB values normalized to a white reference luminance of 100.0,
+    ///
+    /// # Arguments
+    /// * `light` - A reference to an object implementing the [Light] trait, such as [`CieIlluminant`].
+    /// * `filter` - A reference to an object implementing the [Filter] trait, such as `Colorant`.
+    ///
+    /// # Returns
+    /// * `CieLab` - The computed CIELAB color representation for the light and filter combination.
+    pub fn lab(&self, light: &dyn Light, filter: &dyn Filter) -> CieLab {
+        let [xyz, xyzn] = self.xyz_and_xyzn(light, filter);
+        // unwrap OK as we are using only one observer (self) here
+        CieLab::from_xyz(xyz, xyzn).unwrap()
+    }
 
-    */
+    /// Calculates the L*a*b* CIELAB D65 values of a Colorant, using D65 as an illuminant.
+    /// Convenience method for `lab` with D65 illuminant, and using an illuminance value of 100.0.
+    pub fn lab_d65(&self, filter: &dyn Filter) -> CieLab {
+        self.lab(&crate::illuminant::D65, filter)
+    }
+
+    /// Calculates the L*a*b* CIELAB D50 values of a Colorant, using D50 as an illuminant.
+    /// Convenience method for `lab` with D50 illuminant, and using an illuminance value of 100.0.
+    pub fn lab_d50(&self, filter: &dyn Filter) -> CieLab {
+        self.lab(&crate::illuminant::D50, filter)
+    }
+
+    /// Calculates the CIECAM16 color appearance model values for a light source and filter combination.
+    /// This method is used to compute the color appearance of a light source
+    /// when filtered by a colorant or filter, using the CIECAM16 model.
+    /// It returns the CIECAM16 values normalized to a white reference luminance of 100.0.
+    /// # Arguments
+    /// * `light` - A reference to an object implementing the [Light] trait, such as [`CieIlluminant`].
+    /// * `filter` - A reference to an object implementing the [Filter] trait, such as `Colorant`.
+    /// * `vc` - The view conditions to use for the CIECAM16 calculation.
+    /// # Returns
+    /// * `CieCam16` - The computed CIECAM16 color appearance model representation for the light and filter combination.
+    pub fn ciecam16(&self, light: &dyn Light, filter: &dyn Filter, vc: ViewConditions) -> CieCam16 {
+        let [xyz, xyzn] = self.xyz_and_xyzn(light, filter);
+        // unwrap OK as we are using only one observer (self) here
+        CieCam16::from_xyz(xyz, xyzn, vc).unwrap()
+    }
+
+    /// Calculates the CIECAM02 color appearance model values for a light source and filter combination.
+    /// This method is used to compute the color appearance of a light source
+    /// when filtered by a colorant or filter, using the CIECAM02 model.
+    /// It returns the CIECAM02 values normalized to a white reference luminance of 100.0.
+    /// # Arguments
+    /// * `light` - A reference to an object implementing the [Light] trait, such as [`CieIlluminant`].         
+    /// * `filter` - A reference to an object implementing the [Filter] trait, such as `Colorant`.
+    /// * `vc` - The view conditions to use for the CIECAM02 calculation.
+    /// # Returns
+    /// * `CieCam02` - The computed CIECAM02 color appearance model representation for the light and filter combination.
+    pub fn ciecam02(&self, light: &dyn Light, filter: &dyn Filter, vc: ViewConditions) -> CieCam02 {
+        let [xyz, xyzn] = self.xyz_and_xyzn(light, filter);
+        // unwrap OK as we are using only one observer (self) here
+        CieCam02::from_xyz(xyz, xyzn, vc).unwrap()
+    }
+
+    /// Calculates Tristimulus valus, in form of an [XYZ] object of a general spectrum.
+    /// If a reference white is given (rhs), it will copy its tristimulus value, and the spectrum
+    /// is interpreted as a stimulus, being a combination of an illuminant with a colorant.
+    /// If no reference white is given, the spectrum is interpreted as an illuminant.
+    /// This method produces the raw XYZ data, not normalized to 100.0
     pub fn xyz_from_spectrum(&self, spectrum: &Spectrum) -> XYZ {
         let xyz = self.data().data * spectrum.0 * self.data().lumconst;
         XYZ::from_vecs(xyz, self.data().tag)
@@ -262,24 +328,6 @@ impl Observer {
         self.xyz_from_fn(|l| p.slope_at_wavelength(to_wavelength(l, 0.0, 1.0)))
     }
 
-    /// Calculates the L*a*b* CIELAB D50 values of a Colorant, using D65 as an illuminant.
-    /// Accepts a Colorant Spectrum only.
-    /// Returns f64::NAN's otherwise.
-    pub fn lab_d50(&self, filter: &dyn Filter) -> CieLab {
-        let xyz = self.xyz(&CieIlluminant::D50, Some(filter));
-        let xyzn = self.xyz_d50();
-        CieLab::from_xyz(xyz, xyzn).unwrap()
-    }
-
-    /// Calculates the L*a*b* CIELAB D65 values of a Colorant, using D65 as an illuminant.
-    /// Accepts a Colorant Spectrum only.
-    /// Returns f64::NAN's otherwise.
-    pub fn lab_d65(&self, filter: &dyn Filter) -> CieLab {
-        let xyz = self.xyz(&CieIlluminant::D65, Some(filter));
-        let xyzn = self.xyz_d65();
-        CieLab::from_xyz(xyz, xyzn).unwrap()
-    }
-
     /// Calculates the RGB to XYZ matrix, for a particular color space.
     pub fn rgb2xyz(&self, rgbspace: RgbSpace) -> Matrix3<f64> {
         let space = rgbspace.data();
@@ -304,6 +352,10 @@ impl Observer {
         // unwrap: only used with library color spaces
         self.rgb2xyz(rgbspace).try_inverse().unwrap()
     }
+
+    //  pub fn rgb(&self, rgbspace: RgbSpace) -> Matrix3<f64> {
+    //      self.rgb2xyz(rgbspace)
+    //  }
 
     /// Returns the wavelength range (in nanometer) for the _horse shoe_,
     /// the boundary of the area of all physical colors in a chromiticity diagram,
@@ -368,14 +420,16 @@ impl Observer {
     /// and converting the resulting value to RGB values.
     /// ```
     /// use colorimetry::prelude::*;
-    /// let rgb: [u8;3] = Cie1931.xyz_from_colorant_fn(&CieIlluminant::D65, |x|x).rgb(None).clamp().into();
+    /// use colorimetry::rgb::RgbSpace::SRGB;
+    /// let rgb: [u8;3] = Cie1931.xyz_from_colorant_fn(&CieIlluminant::D65, |x|x).rgb(SRGB).clamp().into();
     /// assert_eq!(rgb, [212, 171, 109]);
     /// ```
     /// Linear low pass filter, with a value of 1.0 for a wavelength of 380nm, and a value of 0.0 for 780nm,
     /// and converting the resulting value to RGB values.
     /// ```
     /// use colorimetry::prelude::*;
-    /// let rgb: [u8;3] = Cie1931.xyz_from_colorant_fn(&CieIlluminant::D65, |x|1.0-x).rgb(None).clamp().into();
+    /// use colorimetry::rgb::RgbSpace::SRGB;
+    /// let rgb: [u8;3] = Cie1931.xyz_from_colorant_fn(&CieIlluminant::D65, |x|1.0-x).rgb(SRGB).clamp().into();
     /// assert_eq!(rgb, [158, 202, 237]);
     /// ```
     pub fn xyz_from_colorant_fn(&self, illuminant: &CieIlluminant, f: impl Fn(f64) -> f64) -> XYZ {
@@ -468,23 +522,22 @@ mod obs_test {
             for wavelength in observer.spectral_locus_wavelength_range() {
                 let xyz = observer.xyz_at_wavelength(wavelength).unwrap();
                 for rgbspace in RgbSpace::iter() {
-                    let _rgb = xyz.rgb(Some(rgbspace));
+                    let _rgb = xyz.rgb(rgbspace);
                 }
             }
         }
     }
 
     #[test]
-    // Data from http://www.brucelindbloom.com/index.html?Eqn_RGB_XYZ_Matrix.html
-    // Lindbloom's values are reproduced with an accuracy of 3E-4, which is
-    // a small, but significant difference.  This difference is caused by a difference in the display's white point,
-    // due to wavelength domain differences.  Here we use a domain
-    // from 380 to 780 with a step of 1 nanometer for the spectra, and in
-    // specific for the color matching functions, as recommended by
-    // CIE15:2004, and the color matching functions provided by the CIE in
-    // their online dataset section. The spectra for the primaries are chosen to match the RGB primary values as given by the Color Space specifications.
-    // The white point uses the standard's spectral distribution, as provided by the CIE, clipped to a domain from 380 to 780 nanometer.
-    // See `colorimetry::data::D65`.
+    // Data from http://www.brucelindbloom.com/index.html?Eqn_RGB_XYZ_Matrix.html Lindbloom's values
+    // are reproduced with an accuracy of 3E-4, which is a small, but significant difference.  This
+    // difference is caused by a difference in the display's white point, due to wavelength domain
+    // differences.  Here we use a domain from 380 to 780 with a step of 1 nanometer for the
+    // spectra, and in specific for the color matching functions, as recommended by CIE15:2004, and
+    // the color matching functions provided by the CIE in their online dataset section. The spectra
+    // for the primaries are chosen to match the RGB primary values as given by the Color Space
+    // specifications.  The white point uses the standard's spectral distribution, as provided by
+    // the CIE, clipped to a domain from 380 to 780 nanometer.  See `colorimetry::data::D65`.
     fn test_rgb2xyz_cie1931() {
         let want = nalgebra::Matrix3::new(
             0.4124564, 0.3575761, 0.1804375, 0.2126729, 0.7151522, 0.0721750, 0.0193339, 0.1191920,